+++ src/shared/events/email-address-already-exists.event.ts
